+++ CHANGELOG.md
@@ -11,7 +11,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - **CUMULUS-1686**
   - `ecs_cluster_instance_image_id` is now a *required* variable of the `cumulus` module, instead of optional.
 
+- **CUMULUS-1703**
+  - Remove the unused `forceDownload` option from the `sync-granule` tasks's config
+  - Remove the `@cumulus/ingest/granule.Discover` class
+  - Remove the `@cumulus/ingest/granule.Granule` class
+  - Remove the `@cumulus/ingest/pdr.Discover` class
+  - Remove the `@cumulus/ingest/pdr.Granule` class
+  - Remove the `@cumulus/ingest/parse-pdr.parsePdr` function
+
 ### Added
+
+- **CUMULUS-1703**
+  - Add `@cumulus/aws-client/S3.createBucket` function
+  - Add `@cumulus/aws-client/S3.putFile` function
+  - Add `@cumulus/common/string.isNonEmptyString` function
+  - Add `@cumulus/ingest/FtpProviderClient` class
+  - Add `@cumulus/ingest/HttpProviderClient` class
+  - Add `@cumulus/ingest/S3ProviderClient` class
+  - Add `@cumulus/ingest/SftpProviderClient` class
+  - Add `@cumulus/ingest/providerClientUtils.buildProviderClient` function
+  - Add `@cumulus/ingest/providerClientUtils.fetchTextFile` function
 
 - **CUMULUS-1040**
   - Added `@cumulus/aws-client` package to provide utilities for working with AWS services and the Node.js AWS SDK